# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Fish App Documentation",
        "description": "Documentation",
        "contact": {
            "email": "kanamboma@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost/api",
            "description": "Api Server"
        }
    ],
    "paths": {
...
        "/genres": {
            "get": {
                "tags": [
                    "Genre"
                ],
                "summary": "Retourner la liste des genres",
                "description": "Retourner la liste des genres",
                "operationId": "getAllgenres",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "not found"
                    }
                }
            },
            "post": {
                "tags": [
                    "Genre"
                ],
                "summary": "Ajouter un genre",
                "description": "Ajouter un genre",
                "operationId": "addgenres",
                "requestBody": {
                    "description": "Attribut de la classe genre",
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nom"
                                ],
                                "properties": {
                                    "nom": {
                                        "type": "string",
                                        "example": "XXXX"
                                    },
                                    "illustration": {
                                        "type": "string",
                                        "example": "XXXX"
                                    },
                                    "famile_id": {
                                        "type": "integer",
                                        "example": "XXXX"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "not found"
                    }
                }
            }
        },
        "/genres/{id}": {
            "get": {
                "tags": [
                    "Genre"
                ],
                "summary": "Retourner un genre",
                "description": "Retourner un des genres",
                "operationId": "getAllgenre",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Genre"
                ],
                "summary": "Modifier un genre",
                "description": "Modifier un genre",
                "operationId": "updategenre",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Attribut de la classe genre",
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "nom": {
                                        "type": "string",
                                        "example": "XXXX"
                                    },
                                    "illustration": {
                                        "type": "string",
                                        "example": "XXXX"
                                    },
                                    "famille_id": {
                                        "type": "integer",
                                        "example": "XXXX"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "not found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Genre"
                ],
                "summary": "Supprimer un genre",
                "description": "Supprimer un des genres",
                "operationId": "deletegenre",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "not found"
                    }
                }
            }
        }
    }
}